requirements.txt
easyocr>=1.7.1
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0